refactor(metadata): migrate resources component to v2
