+++ cmd/dogstatsd/subcommands/start/command.go
@@ -52,8 +52,9 @@ import (
 	"github.com/DataDog/datadog-agent/comp/metadata/inventoryagent/inventoryagentimpl"
 	"github.com/DataDog/datadog-agent/comp/metadata/inventoryhost"
 	"github.com/DataDog/datadog-agent/comp/metadata/inventoryhost/inventoryhostimpl"
-	"github.com/DataDog/datadog-agent/comp/metadata/resources"
-	"github.com/DataDog/datadog-agent/comp/metadata/resources/resourcesimpl"
+	resources "github.com/DataDog/datadog-agent/comp/metadata/resources/def"
+	resourcesfx "github.com/DataDog/datadog-agent/comp/metadata/resources/fx"
+	resourcesimpl "github.com/DataDog/datadog-agent/comp/metadata/resources/impl"
 	"github.com/DataDog/datadog-agent/comp/metadata/runner"
 	metadatarunnerimpl "github.com/DataDog/datadog-agent/comp/metadata/runner/runnerimpl"
 	logscompressionfx "github.com/DataDog/datadog-agent/comp/serializer/logscompression/fx"
@@ -160,7 +161,7 @@ func RunDogstatsdFct(cliParams *CLIParams, defaultConfPath string, defaultLogFil
 		}),
 		fx.Supply(resourcesimpl.Disabled()),
 		metadatarunnerimpl.Module(),
-		resourcesimpl.Module(),
+		resourcesfx.Module(),
 		hostimpl.Module(),
 		inventoryagentimpl.Module(),
 		ipcfx.ModuleReadWrite(),
