+++ comp/metadata/host/hostimpl/host_test.go
@@ -21,8 +21,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/core/hostname/hostnameimpl"
 	log "github.com/DataDog/datadog-agent/comp/core/log/def"
 	logmock "github.com/DataDog/datadog-agent/comp/core/log/mock"
-	"github.com/DataDog/datadog-agent/comp/metadata/resources"
-	"github.com/DataDog/datadog-agent/comp/metadata/resources/resourcesimpl"
+	resourcesmock "github.com/DataDog/datadog-agent/comp/metadata/resources/mock"
 	"github.com/DataDog/datadog-agent/pkg/serializer"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
@@ -33,8 +32,8 @@ func TestNewHostProviderDefaultIntervals(t *testing.T) {
 			t,
 			fx.Provide(func() log.Component { return logmock.New(t) }),
 			fx.Provide(func() config.Component { return config.NewMock(t) }),
-			resourcesimpl.MockModule(),
-			fx.Replace(resources.MockParams{Data: nil}),
+			resourcesmock.MockModule(),
+			fx.Replace(resourcesmock.MockParams{Data: nil}),
 			fx.Provide(func() serializer.MetricSerializer { return nil }),
 			hostnameimpl.MockModule(),
 		),
@@ -120,8 +119,8 @@ func TestNewHostProviderIntervalValidation(t *testing.T) {
 					t,
 					fx.Provide(func() log.Component { return logmock.New(t) }),
 					fx.Provide(func() config.Component { return config.NewMockWithOverrides(t, overrides) }),
-					resourcesimpl.MockModule(),
-					fx.Replace(resources.MockParams{Data: nil}),
+					resourcesmock.MockModule(),
+					fx.Replace(resourcesmock.MockParams{Data: nil}),
 					fx.Provide(func() serializer.MetricSerializer { return nil }),
 					hostnameimpl.MockModule(),
 				),
@@ -147,8 +146,8 @@ func TestBackoffWhenEarlyIntervalEqualsCollectionInterval(t *testing.T) {
 	ret := newHostProvider(fxutil.Test[dependencies](t,
 		fx.Provide(func() log.Component { return logmock.New(t) }),
 		fx.Provide(func() config.Component { return config.NewMockWithOverrides(t, overrides) }),
-		resourcesimpl.MockModule(),
-		fx.Replace(resources.MockParams{Data: nil}),
+		resourcesmock.MockModule(),
+		fx.Replace(resourcesmock.MockParams{Data: nil}),
 		fx.Provide(func() serializer.MetricSerializer { return nil }),
 		hostnameimpl.MockModule(),
 	))
@@ -166,8 +165,8 @@ func TestFlareProvider(t *testing.T) {
 			t,
 			fx.Provide(func() log.Component { return logmock.New(t) }),
 			fx.Provide(func() config.Component { return config.NewMock(t) }),
-			resourcesimpl.MockModule(),
-			fx.Replace(resources.MockParams{Data: nil}),
+			resourcesmock.MockModule(),
+			fx.Replace(resourcesmock.MockParams{Data: nil}),
 			fx.Provide(func() serializer.MetricSerializer { return nil }),
 			hostnameimpl.MockModule(),
 		),
@@ -186,8 +185,8 @@ func TestStatusHeaderProvider(t *testing.T) {
 			t,
 			fx.Provide(func() log.Component { return logmock.New(t) }),
 			fx.Provide(func() config.Component { return config.NewMock(t) }),
-			resourcesimpl.MockModule(),
-			fx.Replace(resources.MockParams{Data: nil}),
+			resourcesmock.MockModule(),
+			fx.Replace(resourcesmock.MockParams{Data: nil}),
 			fx.Provide(func() serializer.MetricSerializer { return nil }),
 			hostnameimpl.MockModule(),
 		),
