+++ comp/metadata/resources/def/component.go
