+++ comp/metadata/resources/impl/params.go
@@ -5,11 +5,6 @@
 
 package resourcesimpl
 
-import (
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
-	"go.uber.org/fx"
-)
-
 // Params defines the parameters for the metadata resources component.
 type Params struct {
 	// Disabled determines if the resources payload will be sent. When disabled, the Get method is still available.
@@ -21,10 +16,4 @@ func Disabled() *Params {
 	return &Params{
 		Disabled: true,
 	}
-}
-
-// Module defines the fx options for this component.
-func Module() fxutil.Module {
-	return fxutil.Component(
-		fx.Provide(newResourcesProvider))
 }
